refactor: move client and server side code to src/utils
