+++ packages/slice-machine/server/src/api/common/utils.ts
@@ -3,19 +3,3 @@ export function delay(time: number): Promise<void> {
     setTimeout(resolve, time);
   });
 }
-
-export function preferWroomBase(smApiUrl: string, baseUrl: string): string {
-  try {
-    const urlFromSmJson = new URL(smApiUrl);
-    if (urlFromSmJson.hostname.endsWith(".wroom.io")) {
-      urlFromSmJson.hostname = "wroom.io";
-      return urlFromSmJson.origin;
-    } else if (urlFromSmJson.hostname.endsWith(".wroom.test")) {
-      urlFromSmJson.hostname = "wroom.test";
-      return urlFromSmJson.origin;
-    }
-    return baseUrl;
-  } catch {
-    return baseUrl;
-  }
-}
